SpreadsheetFormatterSelectorEditHateosHttpHandler was SpreadsheetFormatterSelectorEditHateosHttpEntityHandler
